+++ ocrd_models/ocrd_models/ocrd_page_generateds.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 #
-# Generated Sat Jun  6 15:52:02 2020 by generateDS.py version 2.35.20.
+# Generated Sat Jun  6 16:03:42 2020 by generateDS.py version 2.35.20.
 # Python 3.6.9 (default, Apr 18 2020, 01:56:04)  [GCC 8.4.0]
 #
 # Command line options:
@@ -1208,6 +1208,56 @@ class PcGtsType(GeneratedsSuper):
             obj_.original_tagname_ = 'Page'
     def __hash__(self):
         return hash(self.id)
+    def get_AllImagePaths(self, page=True, region=True, line=True, word=True, glyph=True, alternative_images=True):
+        """
+        Get all the image paths referenced in the PAGE-XML document.
+    
+    
+        page (boolean): Get pc:Page level images
+        region (boolean): Get images on pc:*Region level
+        line (boolean) Get images on pc:TextLine level
+        word (boolean) Get images on pc:Word level
+        glyph (boolean) Get images on pc:Glyph level
+        alternative_images (boolean): Get AlternativeImages as well.
+        """
+        from .constants import NAMESPACES, PAGE_REGION_TYPES # pylint: disable=relative-beyond-top-level,import-outside-toplevel
+        from io import StringIO  # pylint: disable=import-outside-toplevel
+        ret = []
+        if page:
+            ret.append(self.get_Page().imageFilename)
+        if alternative_images:
+            # XXX Since we're only interested in the **paths** of the images,
+            # export, parse and xpath are less convoluted than traversing
+            # the generateDS API. Quite possibly not as efficient as could be.
+            sio = StringIO()
+            self.export(
+                    outfile=sio,
+                    level=0,
+                    name_='PcGts',
+                    namespaceprefix_='pc:',
+                    namespacedef_='xmlns:pc="%s" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="%s %s/pagecontent.xsd"' % (
+                        NAMESPACES['page'],
+                        NAMESPACES['page'],
+                        NAMESPACES['page']
+                    ))
+            doc = parsexmlstring_(sio.getvalue())  # pylint: disable=undefined-variable
+            # shortcut
+            if page and region and line and word and glyph:
+                ret += doc.xpath('//page:AlternativeImage/@filename', namespaces=NAMESPACES)
+            else:
+                if page:
+                    ret += doc.xpath('/page:PcGts/page:Page/page:AlternativeImage/@filename', namespaces=NAMESPACES)
+                if region:
+                    for class_ in PAGE_REGION_TYPES:
+                        ret += doc.xpath('//page:%sRegion/page:AlternativeImage/@filename' % class_, namespaces=NAMESPACES)
+                if line:
+                    ret += doc.xpath('//page:TextLine/page:AlternativeImage/@filename', namespaces=NAMESPACES)
+                if word:
+                    ret += doc.xpath('//page:Word/page:AlternativeImage/@filename', namespaces=NAMESPACES)
+                if glyph:
+                    ret += doc.xpath('//page:Glyph/page:AlternativeImage/@filename', namespaces=NAMESPACES)
+    
+        return ret
 # end class PcGtsType