+++ ocrd_models/ocrd_page_user_methods/get_AllImagePaths.py
@@ -1,4 +1,4 @@
-def get_AllImagePaths(self, page=True, region=True, line=True, alternative_images=True):
+def get_AllImagePaths(self, page=True, region=True, line=True, word=True, glyph=True, alternative_images=True):
     """
     Get all the image paths referenced in the PAGE-XML document.
 
@@ -6,17 +6,19 @@ def get_AllImagePaths(self, page=True, region=True, line=True, alternative_image
     page (boolean): Get pc:Page level images
     region (boolean): Get images on pc:*Region level
     line (boolean) Get images on pc:TextLine level
+    word (boolean) Get images on pc:Word level
+    glyph (boolean) Get images on pc:Glyph level
     alternative_images (boolean): Get AlternativeImages as well.
     """
-    from .constants import NAMESPACES
+    from .constants import NAMESPACES, PAGE_REGION_TYPES # pylint: disable=relative-beyond-top-level,import-outside-toplevel
+    from io import StringIO  # pylint: disable=import-outside-toplevel
     ret = []
     if page:
-        return self.get_PageType().get('imageFilename')
+        ret.append(self.get_Page().imageFilename)
     if alternative_images:
         # XXX Since we're only interested in the **paths** of the images,
         # export, parse and xpath are less convoluted than traversing
         # the generateDS API. Quite possibly not as efficient as could be.
-        from io import StringIO
         sio = StringIO()
         self.export(
                 outfile=sio,
@@ -28,11 +30,21 @@ def get_AllImagePaths(self, page=True, region=True, line=True, alternative_image
                     NAMESPACES['page'],
                     NAMESPACES['page']
                 ))
-        # XXX _parsexml parses file with lxml
-        doc = parsexml_(sio.getvalue())
-        if page:
-            for ai in doc.xpath('/page:PcGts/page:Page/page:AlternativeImage', namespaces=NAMESPACES):
-                print(ai)
+        doc = parsexmlstring_(sio.getvalue())  # pylint: disable=undefined-variable
+        # shortcut
+        if page and region and line and word and glyph:
+            ret += doc.xpath('//page:AlternativeImage/@filename', namespaces=NAMESPACES)
+        else:
+            if page:
+                ret += doc.xpath('/page:PcGts/page:Page/page:AlternativeImage/@filename', namespaces=NAMESPACES)
+            if region:
+                for class_ in PAGE_REGION_TYPES:
+                    ret += doc.xpath('//page:%sRegion/page:AlternativeImage/@filename' % class_, namespaces=NAMESPACES)
+            if line:
+                ret += doc.xpath('//page:TextLine/page:AlternativeImage/@filename', namespaces=NAMESPACES)
+            if word:
+                ret += doc.xpath('//page:Word/page:AlternativeImage/@filename', namespaces=NAMESPACES)
+            if glyph:
+                ret += doc.xpath('//page:Glyph/page:AlternativeImage/@filename', namespaces=NAMESPACES)
 
-
-
+    return ret